Pong 0
Restarting the Pong Project, and I want to do multiple updates as I go, and note what changed or what I learned.

With this, the big thing I got stuck on was trying to debug why love.graphics.printf wasn't working properly, and the issue was just that WINDOW_WIDTH/HEIGHT were in the wrong order, so every function that took them in as an argument was mixing them up, causing weird issues with where the text would appear.

It's also important to note, all functions seem to require 'end' with lua. A lot of my issues getting it to run was just missing an 'end' somewhere.